add camera rail mode to Crash 1 zones (#163)
* add wavey animation to the box links

* help text consistency + change opengl error msg format

* optimize collision renderer to not wreck GC

* use names for culling mode text

* minor changes to some eid lookup

* organize lerp functions

* fully disable zbuffer for text renderer

* add "anchor mode" toggle to crash 1 zone viewer

* of course this method doesn't work, why would it

* don't lock anchor zone to just the master zone

* fix errors in viewer input handling

* optimize world viewer for untextured mode

* use sort list in anchor mode by default

* move slst decode code to API

* fix bug with camera z rot value

* fix bugs with neighbor cameras

* add a detach mode

* bump rev instead

+++ Crash/Formats/Crash Formats/Scenery/OldSceneryPolygon.cs
@@ -32,18 +32,18 @@ namespace Crash
                 throw new ArgumentOutOfRangeException(nameof(vertexc));
             if (modelstruct < 0 || modelstruct > 0xFFF)
                 throw new ArgumentOutOfRangeException(nameof(modelstruct));
-            VertexA = vertexa;
-            VertexB = vertexb;
-            VertexC = vertexc;
+            VertexA = (short)vertexa;
+            VertexB = (short)vertexb;
+            VertexC = (short)vertexc;
             ModelStruct = modelstruct;
             Page = page;
             Anim0 = anim0;
             Unknown = unknown;
         }
 
-        public int VertexA { get; set; }
-        public int VertexB { get; set; }
-        public int VertexC { get; set; }
+        public short VertexA { get; set; }
+        public short VertexB { get; set; }
+        public short VertexC { get; set; }
         public int ModelStruct { get; }
         public byte Page { get; }
         public byte Anim0 { get; }
